Move tests about downloading sample cases to online-judge-tools/api-client repository
See also https://github.com/online-judge-tools/api-client/pull/132
